+++ app/src/main/java/com/jin/honey/feature/cart/ui/CartScreen.kt
@@ -1,4 +1,4 @@
-package com.jin.honey.feature.order.ui.content.cart
+package com.jin.honey.feature.cart.ui
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
@@ -23,17 +23,14 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.compose.ui.util.trace
 import com.jin.honey.R
 import com.jin.honey.feature.cart.domain.model.Cart
 import com.jin.honey.feature.cart.domain.model.CartKey
-import com.jin.honey.feature.order.ui.content.cart.content.CartContent
-import com.jin.honey.feature.order.ui.content.cart.content.CartHeader
-import com.jin.honey.feature.order.ui.content.cart.content.CartOptionModifyBottomSheet
-import com.jin.honey.ui.theme.HoneyTheme
+import com.jin.honey.feature.cart.ui.content.CartContent
+import com.jin.honey.feature.cart.ui.content.CartHeader
+import com.jin.honey.feature.cart.ui.content.CartOptionModifyBottomSheet
 import com.jin.honey.ui.theme.OrderDetailBoxBorderColor
 
 @Composable
