+++ app/src/main/java/com/jin/honey/feature/cart/ui/content/CartHeader.kt
@@ -1,10 +1,8 @@
-package com.jin.honey.feature.order.ui.content.cart.content
+package com.jin.honey.feature.cart.ui.content
 
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ShoppingCart
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
