+++ app/src/main/java/com/jin/honey/feature/order/ui/OrderScreen.kt
@@ -17,8 +17,8 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.jin.honey.R
-import com.jin.honey.feature.order.ui.content.cart.CartScreen
-import com.jin.honey.feature.order.ui.content.orderhistory.OrderHistoryScreen
+import com.jin.honey.feature.cart.ui.CartScreen
+import com.jin.honey.feature.order.ui.content.OrderHistoryScreen
 import com.jin.honey.feature.ui.state.DbState
 import com.jin.honey.feature.ui.state.UiState
