+++ app/src/main/java/com/jin/honey/feature/order/ui/content/OrderHistoryScreen.kt
@@ -1,4 +1,4 @@
-package com.jin.honey.feature.order.ui.content.orderhistory
+package com.jin.honey.feature.order.ui.content
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
